requirements.txt
discord.py
requests
python-dotenv
SQLAlchemy==2.0.23
openai